@@ -26,26 +26,30 @@
   "license": "ISC",
   "repository": "{{repo}}",
   "dependencies": {
-    "gsap": "^1.19.1"{{#is framework 'bigwheel'}},
+    "gsap": "^1.19.1",
+    "howler": "^2.0.4"{{#is framework 'bigwheel'}},
     "bigwheel": "^3.0.0",
     "domready": "^1.0.8",
     "bw-analytics": "^1.1.1",
     "handlebars": "^2.0.0",
     "domify": "^1.4.0"{{/is}}{{#is framework 'react'}},
-    "babel-polyfill": "^6.7.4",
+    "babel-polyfill": "^6.23.0",
     "domready": "^1.0.8",
-    "mobile-detect": "^1.3.3",
-    "preloader": "^2.1.0",
-    "react": "^15.3.2",
-    "react-addons-perf": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-svg-inline": "^1.2.0",
+    "fullscreen-handler": "^0.0.1",
+    "mobile-detect": "^1.3.6",
+    "preloader": "^4.0.1",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
+    "react-addons-perf": "^15.4.2",
+    "react-background-video-player": "^0.2.1",
+    "react-dom": "^15.6.1",
+    "react-svg-inline": "^2.0.0",
     "react-redux": "^4.4.4",
     "react-router": "^2.1.1",
     "react-router-redux": "^4.0.2",
-    "react-transition-group-plus": "^0.4.0",
+    "react-transition-group-plus": "^0.4.2",
     "redux": "^3.4.0",
-    "stats.js": "0.16.0"{{/is}}
+    "stats.js": "0.17.0"{{/is}}
   },
   "devDependencies": {
     "budo": "^10.0.3",
