@@ -1,13 +1,14 @@
 'use strict';
-var MobileDetect = require('mobile-detect');
-var ua = navigator.userAgent;
-var md = new MobileDetect(ua);
-var utilOS = require('./util-os');
-var utilBrowser = require('./util-browser');
-var env = process.env.NODE_ENV || 'development';
+import MobileDetect from 'mobile-detect';
+import utilOS from './util-os';
+import utilBrowser from './util-browser';
 
-var checkDevice = function() {
-  var device = 'desktop';
+const ua = navigator.userAgent;
+const md = new MobileDetect(ua);
+const env = process.env.NODE_ENV || 'development';
+
+const checkDevice = function() {
+  let device = 'desktop';
   if (md.mobile() && md.phone()) {
     device = 'phone';
   } else if (md.mobile() && md.tablet()) {
@@ -16,15 +17,15 @@ var checkDevice = function() {
   return device;
 };
 
-var checkVendor = function() {
+const checkVendor = function() {
   return (navigator.vendor) ? navigator.vendor.toLowerCase() : '';
 };
 
-var checkBrowser = function() {
-  var browser = 'unknown';
-  var uaLower = ua.toLowerCase();
-  var msie = uaLower.indexOf('msie') >= 0;
-  var trident = uaLower.indexOf('trident/') >= 0;
+const checkBrowser = function() {
+  let browser = 'unknown';
+  const uaLower = ua.toLowerCase();
+  const msie = uaLower.indexOf('msie') >= 0;
+  const trident = uaLower.indexOf('trident/') >= 0;
 
   if (msie || trident) {
     browser = 'ie';
@@ -42,14 +43,14 @@ var checkBrowser = function() {
   return browser;
 };
 
-var checkDevicePixelRatio = function() {
-  var pxlRatio = window.devicePixelRatio;
+const checkDevicePixelRatio = function() {
+  let pxlRatio = window.devicePixelRatio;
   if (utilOS.os() === 'iOS' && window.innerWidth >= 375 && window.devicePixelRatio < 3) pxlRatio = 3;
   return pxlRatio;
 };
 
-var checkManufacturer = function() {
-  var man = 'unknown';
+const checkManufacturer = function() {
+  let man = 'unknown';
   if (md.phone()) {
     man = md.phone();
   } else if (md.tablet()) {
@@ -58,10 +59,12 @@ var checkManufacturer = function() {
   return man.toLowerCase();
 };
 
-var getClasses = function() {
-  var classes = [checkDevice(), 'x' + checkDevicePixelRatio(), checkBrowser(), 'v' + utilBrowser.checkVersion(), (utilOS.os()).replace(/\s/g, '_').toLocaleLowerCase()];
+const getClasses = function() {
+  const classes = [checkDevice(), 'x' + checkDevicePixelRatio(), checkBrowser(), 'v' + utilBrowser.checkVersion(), (utilOS.os()).replace(/\s/g, '_').toLocaleLowerCase()];
   if (md.mobile()) classes.push(checkManufacturer());
-  return classes.filter(function(cur) { return !!cur; });
+  return classes.filter(function(cur) {
+    return !!cur;
+  });
 };
 
 module.exports = {
@@ -91,14 +94,14 @@ module.exports = {
   md: md,
   get orientation() {
     if (window.screen) {
-      var orientation = window.screen.orientation || window.screen.mozOrientation || window.screen.msOrientation;
+      const orientation = window.screen.orientation || window.screen.mozOrientation || window.screen.msOrientation;
       if (orientation && orientation.type) {
         return orientation.type.split('-', 1)[0];
       }
     }
-    var w = Math.max(document.documentElement.clientWidth, window.innerWidth || 0);
-    var h = Math.max(document.documentElement.clientHeight, window.innerHeight || 0);
-    var aspectRatio = w / h;
+    const w = Math.max(document.documentElement.clientWidth, window.innerWidth || 0);
+    const h = Math.max(document.documentElement.clientHeight, window.innerHeight || 0);
+    const aspectRatio = w / h;
     if (aspectRatio < 1) {
       return 'portrait';
     } else {
