@@ -5,21 +5,18 @@ import {Router, Route, IndexRoute, IndexRedirect, useRouterHistory, Redirect} fr
 import {{#if pushState}}createBrowserHistory{{else}}createHashHistory{{/if}} from 'history/lib/{{#if pushState}}createBrowserHistory{{else}}createHashHistory{{/if}}'
 import { syncHistoryWithStore, routerReducer } from 'react-router-redux';
 import { Provider } from 'react-redux';
-import detect from '../util/detect';
 
+import detect from '../util/detect';
 import store from '../store';
-import App from '../sections/App{{#if sectionNames}}/App{{/if}}';
-import Landing from '../sections/Landing{{#if sectionNames}}/Landing{{/if}}';
+
+import routeKeys from './route-keys';
+import App from '../sections/App';
+import Landing from '../sections/Landing';
 
 const history = syncHistoryWithStore(useRouterHistory({{#if pushState}}createBrowserHistory{{else}}createHashHistory{{/if}})({ basename: process.env.BASENAME }), store);
 
 let Test;
 let TestRoutes;
-if (process.env.NODE_ENV === 'development') {
-  window.Perf = require('react-addons-perf');
-  Test = require('../test/index').default;
-  TestRoutes = require('../test/routes').default;
-}
 
 export default function() {
   const container = document.createElement('div');
@@ -27,15 +24,28 @@ export default function() {
   document.body.appendChild(container);
   document.body.className = [...document.body.className.split(' '), ...detect.classes].join(' ');
 
+  (process.env.NODE_ENV === 'development') && setDevTools();
+
   render((
     <Provider store={store}>
-      <Router history={history}>
-        <Route path="/" component={App}>
-          <IndexRoute component={Landing}/>
-          {TestRoutes && <Route path="/test" component={Test}>{TestRoutes}</Route>}
-          <Redirect path="*" to="/"/>
-        </Route>
-      </Router>
-    </Provider>
-  ), container);
+      <Router
+          history={history}
+          onUpdate={handleRouteChange}
+        >
+          <Route component={App}>
+            <Route path={routeKeys.Landing} component={Landing}/>
+            {TestRoutes && <Route path={routeKeys.Test} component={Test}>{TestRoutes}</Route>}
+            <Redirect path="*" to={routeKeys.Landing}/>
+          </Route>
+        </Router>
+      </Provider>
+    ), container);
+  }
+
+function setDevTools() {
+  Test = require('../test/index').default;
+  TestRoutes = require('../test/routes').default;
+  window.Perf = require('react-addons-perf');
 }
+
+function handleRouteChange() {}
