@@ -1,5 +1,6 @@
 export default {
   SET_READY: 'SET_READY',
   SET_PROGRESS: 'SET_PROGRESS',
-  SET_ASSETS: 'SET_ASSETS'
-}
+  SET_ASSETS: 'SET_ASSETS',
+  SET_WINDOW_SIZE: 'SET_WINDOW_SIZE',
+};
