@@ -5,8 +5,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import TransitionGroup from 'react-transition-group-plus';
-import Preloader from '../../components/Preloader{{#if sectionNames}}/Preloader{{/if}}';
-import { setReady, setProgress } from '../../sections/App/actions';
+import Preloader from '../../../components/Preloader/Preloader';
+import { setReady } from '../../../store/actions/app';
+import { setProgress } from '../../../store/actions/preloader';
 
 class Test extends React.Component {
   componentWillUnmount() {
@@ -21,7 +22,7 @@ class Test extends React.Component {
           !this.props.ready &&
           <Preloader
             {...this.props}
-            assetsList={this.props.assets}
+            assets={this.props.assets}
             setProgress={this.props.setProgress}
             setReady={this.props.setReady}
           />
@@ -32,7 +33,7 @@ class Test extends React.Component {
 }
 
 const mapStateToProps = state => ({
-  assets: state.assets,
+  assets: state.preloader.assets,
   ready: state.ready
 });
 
